clean code + riverpod state management
