+++ pubspec.yaml
@@ -37,6 +37,7 @@ dependencies:
   cupertino_icons: ^1.0.2
   provider: ^6.0.5
   google_fonts: ^4.0.4
+  flutter_riverpod: ^2.3.6
 
 dev_dependencies:
   flutter_test:
